Changes the args so that individual input files can be specified...
instead of taking a whole directory.

Sometimes its nice to just not have to process, or see the output of,
a whole directory full of loadouts. So I've changed the arguments so
you need to specify each individual input file. A whole directory can
still be passed by doing `in/*` and letting the shell do glob
expansion. So this gives more flexibility.